Nit: change wording on use of the subtle crate.
It's not that the crate is relying on `subtle` to perform constant-time
arithmetic, because the arithmetic implementations are done in this crate and
it's the fact that they're implemented in a constant-time way that's important.
Using subtle helps with constant-time operations but it is not enough, and
changing the wording of that sentence clarifies that.

(This is a nit).