Implement curve Generator
Here we implement set a curve generator
function. This is done by converting bytes
into random curve points, and iterating
until one is found which satisfies the
properties of 'is_prime_order', fucntion.

Additionally, the fixed generator is removed
from the affine IMPL block. However, the
fixed point is kept, for reasons explained
in the comments.