DHKE should return canonical EC point
The shared secret used on the diffie-hellman protocol will be an
elliptic curve point. It will be up to the user whether he wants to use
the 32-byte scalar representation of the Y coordinate or perform
operations with the shared secret.

The motivation is to not enforce the user to reconstruct the point in
case he need to perform ECC operations.