Fix ZIP 216 bug
Since we're not in prod yet, we don't need to be backwards compatible
with the previous encoding and simply accept the next one.

The solution prevents a point of order 2 (0, -1) to be accepted as a
valid point.

For more details, see: https://zips.z.cash/zip-0216

Special thanks to Zcash and @str4d for pinging and filing an issue in
regards of that.

Resolves:  #82